Verification of JAR signature moved from manifest to jarutil.
Refactored signing tests, updated README for tests, cleaned up test files.
Added test for verification of manifest checksums.
