+++ tests/data/cli-sign-and-verify.jar
